backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
dspy>=2.5.0
anthropic>=0.40.0
elevenlabs>=1.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
aiofiles>=24.0.0
# Auth & Billing
PyJWT>=2.8.0
google-auth>=2.29.0
stripe>=10.0.0
